+++ Q-municate_app/src/main/java/com/quickblox/q_municate/ui/adapters/chats/DialogsListAdapter.java
@@ -70,7 +70,7 @@ public class DialogsListAdapter extends BaseListAdapter<Dialog> {
             displayGroupPhotoImage(dialog.getPhoto(), viewHolder.avatarImageView);
         }
 
-        List<Integer> dialogOccupantsIdsList = ChatUtils.getIdsFromDialogOccupantsList(dialogOccupantsList);
+        List<Long> dialogOccupantsIdsList = ChatUtils.getIdsFromDialogOccupantsList(dialogOccupantsList);
         long unreadMessages = dataManager.getMessageDataManager().getCountUnreadMessages(dialogOccupantsIdsList, currentUser.getId());
         long unreadDialogNotifications = dataManager.getDialogNotificationDataManager().getCountUnreadDialogNotifications(dialogOccupantsIdsList, currentUser.getId());
